# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserController_deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/searchuser": {
      "post": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userSearchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/adduser": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/edituser": {
      "patch": {
        "operationId": "UserController_edit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/role": {
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/role/{name}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/role/{id}": {
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/type": {
      "post": {
        "operationId": "TypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TypeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/type/{id}": {
      "patch": {
        "operationId": "TypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/transaction": {
      "post": {
        "operationId": "TransactionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TransactionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/transaction/{id}": {
      "get": {
        "operationId": "TransactionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "TransactionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TransactionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Mobile Backoffice",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "userSearchDto": {
        "type": "object",
        "properties": {
          "search_value": {
            "type": "string"
          },
          "page_number": {
            "type": "number"
          },
          "page_size": {
            "type": "number"
          },
          "order_by": {
            "type": "string"
          },
          "sort_column": {
            "type": "string"
          }
        },
        "required": [
          "search_value",
          "page_number",
          "page_size",
          "order_by",
          "sort_column"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role_id": {
            "type": "number"
          },
          "branch_id": {
            "type": "number"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "nick_name": {
            "type": "string"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "example": "new Date().toISOString()"
          },
          "age": {
            "type": "number"
          },
          "sex": {
            "type": "string",
            "example": "Male || Female"
          },
          "phone_no": {
            "type": "string",
            "example": "084xxx9514"
          },
          "identity_id": {
            "type": "string",
            "example": "11015xxxxx319"
          },
          "email": {
            "type": "string"
          },
          "salary_base": {
            "type": "number"
          }
        },
        "required": [
          "username",
          "password",
          "role_id",
          "branch_id",
          "first_name",
          "last_name",
          "nick_name",
          "date_of_birth",
          "age",
          "sex",
          "phone_no",
          "identity_id",
          "email",
          "salary_base"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "userId ex: 1​"
          },
          "role_id": {
            "type": "number"
          },
          "branch_id": {
            "type": "number"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "nick_name": {
            "type": "string"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "example": "ใช้ new Date().toISOString()"
          },
          "age": {
            "type": "number"
          },
          "sex": {
            "type": "string"
          },
          "phone_no": {
            "type": "string"
          },
          "identity_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "salary_base": {
            "type": "number"
          },
          "lastest_login": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "list": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "list"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "list": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number"
          },
          "total_price_sold": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "number"
          }
        },
        "required": [
          "product_id",
          "total_price_sold",
          "type"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "number"
          },
          "total_price_sold": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "number"
          }
        }
      }
    }
  }
}